requirements.txt
djangorestframework
pika
python-decouple
redis
django-redis
Jinja2